update to reqnroll as specflow/trisentrics dropped support
