+++ .gitignore
@@ -363,5 +363,5 @@ MigrationBackup/
 # Fody - auto-generated XML schema
 FodyWeavers.xsd
 
-# Specflow feature.cs files
+# Reqnroll feature.cs files
 *.feature.cs
