+++ Reqnroll.AutofacServiceProvider.Tests/Features/DependencyInjectionPlugin.feature
