+++ Reqnroll.AutofacServiceProvider.Tests/Features/UnitTestRuntimeProvider.feature
@@ -1,0 +1,6 @@
+﻿Feature: UnitTestRuntimeProvider
+    Issue #73: Unable to resolve IUnitTestRuntimeProvider
+    https://github.com/solidtoken/Reqnroll.DependencyInjection/issues/73
+
+Scenario: Assert IUnitTestRuntimeProvider is available
+	Then verify that IUnitTestRuntimeProvider is correctly injected
