+++ Reqnroll.AutofacServiceProvider.Tests/Support/ICalculator.cs
@@ -1,0 +1,11 @@
+﻿namespace NativeWaves.Reqnroll.AutofacServiceProvider.Tests.Support
+{
+    public interface ICalculator
+    {
+        void Enter(int operand);
+        void Add();
+        void Multiply();
+        int Result { get; }
+        int Size { get; }
+    }
+}
