+++ Reqnroll.AutofacServiceProvider/IServiceCollectionFinder.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace NativeWaves.Reqnroll.AutofacServiceProvider
+{
+    public interface IServiceCollectionFinder
+    {
+        IServiceCollection GetServiceCollection();
+    }
+}
