--- SpecFlow.AutofacServiceProvider.Tests/Features/UnitTestRuntimeProvider.feature
@@ -1,6 +1,0 @@
-﻿Feature: UnitTestRuntimeProvider
-    Issue #73: Unable to resolve IUnitTestRuntimeProvider
-    https://github.com/solidtoken/SpecFlow.DependencyInjection/issues/73
-
-Scenario: Assert IUnitTestRuntimeProvider is available
-	Then verify that IUnitTestRuntimeProvider is correctly injected
