--- SpecFlow.AutofacServiceProvider.Tests/SpecFlow.AutofacServiceProvider.Tests.csproj
@@ -1,28 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0</TargetFrameworks>
-    <IsPublishable>false</IsPublishable>
-    <IsPackable>false</IsPackable>
-    <Product>NativeWaves.SpecFlow.AutofacServiceProvider</Product>
-    <RootNamespace>NativeWaves.SpecFlow.AutofacServiceProvider.Tests</RootNamespace>
-    <AssemblyName>NativeWaves.SpecFlow.AutofacServiceProvider.Tests</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="SpecFlow.xUnit" Version="[3.9.74]" />
-    <!--<PackageReference Include="SpecRun.SpecFlow" Version="[3.9.31]" />-->
-    <!-- i don't know why we need nunit on the server?!??! -->
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="[17.7.2]" />
-    <PackageReference Include="xunit" Version="[2.5.1]" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="[2.5.1]" PrivateAssets="all" />
-    <PackageReference Include="FluentAssertions" Version="[6.8.0]" PrivateAssets="all" />
-    <PackageReference Include="coverlet.collector" Version="[3.2.0]" PrivateAssets="all" />
-    <PackageReference Include="GitVersion.MsBuild" Version="[5.11.1]" PrivateAssets="all" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SpecFlow.AutofacServiceProvider\SpecFlow.AutofacServiceProvider.csproj" />
-  </ItemGroup>
-
-</Project>
