--- SpecFlow.AutofacServiceProvider.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.31903.286
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpecFlow.AutofacServiceProvider", "SpecFlow.AutofacServiceProvider\SpecFlow.AutofacServiceProvider.csproj", "{CACF906A-7230-4E37-984C-739AB1620995}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F6B4EBE9-9995-403D-BC3F-AD26CD6E960F}"
-	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
-		.gitignore = .gitignore
-		.circleci\config.yml = .circleci\config.yml
-		GitVersion.yml = GitVersion.yml
-		LICENSE = LICENSE
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpecFlow.AutofacServiceProvider.Tests", "SpecFlow.AutofacServiceProvider.Tests\SpecFlow.AutofacServiceProvider.Tests.csproj", "{BC1B8D53-80B7-4CD5-B4D0-2A697FBAF97F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CACF906A-7230-4E37-984C-739AB1620995}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CACF906A-7230-4E37-984C-739AB1620995}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CACF906A-7230-4E37-984C-739AB1620995}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CACF906A-7230-4E37-984C-739AB1620995}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BC1B8D53-80B7-4CD5-B4D0-2A697FBAF97F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BC1B8D53-80B7-4CD5-B4D0-2A697FBAF97F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BC1B8D53-80B7-4CD5-B4D0-2A697FBAF97F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BC1B8D53-80B7-4CD5-B4D0-2A697FBAF97F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CF1B75FE-E67E-486F-9B92-BDBAD9A8B86A}
-	EndGlobalSection
-EndGlobal
