--- SpecFlow.AutofacServiceProvider/IServiceCollectionFinder.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace NativeWaves.SpecFlow.AutofacServiceProvider
-{
-    public interface IServiceCollectionFinder
-    {
-        IServiceCollection GetServiceCollection();
-    }
-}
